requirements.txt
# Core Backend Dependencies
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart
python-dotenv>=1.0.0

# LangChain Core
langchain>=0.0.350
langchain-community>=0.0.10
langchain-core>=0.1.0

# LLM Providers
langchain-groq>=0.0.6  # Groq for fast inference

# Embedding Providers
# Local embeddings (recommended - no API key needed)
sentence-transformers>=2.2.0
torch>=2.0.0  # Required by sentence-transformers

# Google Gemini embeddings (optional - needs API key)
google-genai>=0.5.0
langchain-google-genai>=2.0.0

# Vector Store
faiss-cpu>=1.7.4  # Use faiss-gpu if you have CUDA
numpy>=1.24.0

# Document Processing
pypdf>=3.0.0
pymupdf>=1.23.0  # Better PDF text extraction
pdfplumber>=0.9.0  # Advanced PDF processing

# Data Processing
pandas>=2.0.0

# Optional: Redis for caching
redis>=4.5.0

# Development Dependencies (optional)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0